Fix uninitialized workflow input in 'simpleflow standalone'
Bug detected by @ybastide and mentionned in #106, but actually providing
a blank input is not incorrect in SWF. In simpleflow case we can just
pass 'null' or a blank string (this should be refined later, this is
obviously not the same regarding parameters passed to the workflow).
In the end the 'input' parameter in StartWorkflowExecution is not even
required.